pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Juma Anton",email = "jumaantony840@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai (>=1.23.0,<2.0.0)",
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]

[tool.poetry]
packages = [{include = "chatbot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
